refactor: 테마 설정 로직 분리 및 로고 이미지 통일
* refactor: `authStore`의 테마 적용 로직(`applyTheme`)을 `settingsStore`로 이관
* chore: 로그인 및 프로필 관련 컴포넌트의 로고 이미지를 `logo_green1.png`로 통일
* fix: `index.html` 뷰포트 메타 태그에 모바일 화면 확대 방지(`user-scalable=0` 등) 설정 추가
* fix: `settings.js` 폰트 사이즈 및 고대비 설정에 타입 안정성(소문자 변환 및 boolean 캐스팅) 보장 로직 추가
* fix: `instance.js` 리프레시 토큰 만료에 따른 강제 로그아웃 시 라우터 이동 비동기(`await`) 처리
* chore: `ProfileEdit.vue` 내 불필요한 `axios` 임포트 제거 및 기본 프로필 이미지 에셋 갱신
